requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
numpy>=2.0.0
markdown>=3.6

langchain>=1.2.0
langchain-core>=1.2.0
langchain-community>=0.4.1
langchain-openai>=1.1.6
langchain-text-splitters>=1.1.0

openai>=2.0.0
httpx>=0.27.0
